backend/requirements-core.txt
# Core dependencies - install first for better Docker layer caching
# These are typically slower to install and change less frequently

# Web framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and settings
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0

# Database and migrations
sqlalchemy==2.0.23
alembic==1.12.1

# Async database drivers
aiosqlite==0.18.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# OCR & PDF Processing
marker-pdf==1.9.2
opencv-python-headless>=4.8.0
Pillow>=10.0.0
pycocotools>=2.0.6

# Vector Database
qdrant-client>=1.10.1,<2.0.0
langchain-qdrant==0.1.4

# ML & Text Processing
cohere>=5.0.0
pandas>=1.5.0
rank_bm25

# PyTorch with CUDA support
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0
